Update Room and Customer operations

+++ src/main/java/gr/majestic/reservations/service/RoomService.java
@@ -2,5 +2,5 @@ package gr.majestic.reservations.service;
 
 import gr.majestic.reservations.model.Room;
 
-public interface RoomService extends GeneralService<Room, Integer> {
+public interface RoomService extends GeneralService<Room, Long> {
 }
